+++ coverage-tests/configuration/override-tests.js
@@ -63,26 +63,10 @@ module.exports = {
     'check frame fully with vg': {skip: true}, //  stale element reference
     'check frame fully with css stitching': {skip: true}, // diffs
     'check frame after manual switch to frame with vg classic': {skip: true}, // java.lang.IllegalArgumentException: checkTasks == 0
-    'should send dom and location when check window': {skipEmit: true},
-    'should send dom and location when check window with vg': {skipEmit: true},
-    'should send dom and location when check window fully': {skipEmit: true},
-    'should send dom and location when check window fully with vg': {skipEmit: true},
-    'should send dom and location when check frame': {skipEmit: true},
-    'should send dom and location when check frame with vg': {skipEmit: true},
-    'should send dom and location when check frame fully': {skipEmit: true},
-    'should send dom and location when check frame fully with vg': {skipEmit: true},
-    'should send dom and location when check region by selector': {skipEmit: true},
-    'should send dom and location when check region by selector with vg': {skipEmit: true},
-    'should send dom and location when check region by selector fully': {skipEmit: true},
-    'should send dom and location when check region by selector fully with vg': {skipEmit: true},
-    'should send dom and location when check region by selector in frame': {skipEmit: true},
-    'should send dom and location when check region by selector in frame with vg': {skipEmit: true},
-    'should send dom and location when check region by selector with custom scroll root': {skipEmit: true},
-    'should send dom and location when check region by selector with custom scroll root with vg': {skipEmit: true},
-    'should send dom and location when check region by selector fully with custom scroll root': {skipEmit: true},
-    'should send dom and location when check region by selector fully with custom scroll root with vg': {skipEmit: true},
-    'should send dom of version 10': {skipEmit: true},
     'appium android check window': {skip: true}, //wrong ignore region ticket 2396
     'appium android check region with ignore region': {skip: true}, //wrong ignore region region ticket 2396
     'appium iOS check region with ignore region': {skip: true}, //wrong ignore region region ticket 2396
+
+    'should send dom and location when check frame fully with vg': {skip: true}, // not supported by ufg
+    'should send dom and location when check region by selector in frame with vg': {skip: true}, // not supported by ufg
  }